SignUpPage error handling + confirm password field
Added a confirm password field, as well as error handling for any sign up errors, such as the password not being long enough, the terms and services checkbox not being checked, email not being in use, or full name not being typed in. Added an error_handling folder with an auth-errors file used to handle the errors that arise when trying to register a user.